exclude the cordoned nodes when calc node allocatable
