+++ pkg/spoke/managedcluster/joining_controller_test.go
@@ -10,13 +10,8 @@ import (
 	clusterv1 "open-cluster-management.io/api/cluster/v1"
 	testinghelpers "open-cluster-management.io/registration/pkg/helpers/testing"
 
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	kubeinformers "k8s.io/client-go/informers"
-	kubefake "k8s.io/client-go/kubernetes/fake"
-	kubeversion "k8s.io/client-go/pkg/version"
 	clienttesting "k8s.io/client-go/testing"
 )
 
@@ -24,7 +19,6 @@ func TestSyncManagedCluster(t *testing.T) {
 	cases := []struct {
 		name            string
 		startingObjects []runtime.Object
-		nodes           []runtime.Object
 		validateActions func(t *testing.T, actions []clienttesting.Action)
 		expectedErr     string
 	}{
@@ -42,9 +36,6 @@ func TestSyncManagedCluster(t *testing.T) {
 		{
 			name:            "sync an accepted managed cluster",
 			startingObjects: []runtime.Object{testinghelpers.NewAcceptedManagedCluster()},
-			nodes: []runtime.Object{
-				testinghelpers.NewNode("testnode1", testinghelpers.NewResourceList(32, 64), testinghelpers.NewResourceList(16, 32)),
-			},
 			validateActions: func(t *testing.T, actions []clienttesting.Action) {
 				expectedCondition := metav1.Condition{
 					Type:    clusterv1.ManagedClusterConditionJoined,
@@ -52,110 +43,9 @@ func TestSyncManagedCluster(t *testing.T) {
 					Reason:  "ManagedClusterJoined",
 					Message: "Managed cluster joined",
 				}
-				expectedStatus := clusterv1.ManagedClusterStatus{
-					Version: clusterv1.ManagedClusterVersion{
-						Kubernetes: kubeversion.Get().GitVersion,
-					},
-					Capacity: clusterv1.ResourceList{
-						clusterv1.ResourceCPU:    *resource.NewQuantity(int64(32), resource.DecimalExponent),
-						clusterv1.ResourceMemory: *resource.NewQuantity(int64(1024*1024*64), resource.BinarySI),
-					},
-					Allocatable: clusterv1.ResourceList{
-						clusterv1.ResourceCPU:    *resource.NewQuantity(int64(16), resource.DecimalExponent),
-						clusterv1.ResourceMemory: *resource.NewQuantity(int64(1024*1024*32), resource.BinarySI),
-					},
-				}
 				testinghelpers.AssertActions(t, actions, "get", "update")
 				actual := actions[1].(clienttesting.UpdateActionImpl).Object
 				testinghelpers.AssertManagedClusterCondition(t, actual.(*clusterv1.ManagedCluster).Status.Conditions, expectedCondition)
-				testinghelpers.AssertManagedClusterStatus(t, actual.(*clusterv1.ManagedCluster).Status, expectedStatus)
-			},
-		},
-		{
-			name: "sync a joined managed cluster without status change",
-			startingObjects: []runtime.Object{
-				testinghelpers.NewManagedClusterWithStatus(testinghelpers.NewResourceList(32, 64), testinghelpers.NewResourceList(16, 32)),
-			},
-			nodes: []runtime.Object{
-				testinghelpers.NewNode("testnode1", testinghelpers.NewResourceList(32, 64), testinghelpers.NewResourceList(16, 32)),
-			},
-			validateActions: func(t *testing.T, actions []clienttesting.Action) {
-				testinghelpers.AssertActions(t, actions, "get")
-			},
-		},
-		{
-			name:            "sync a joined managed cluster with status change",
-			startingObjects: []runtime.Object{testinghelpers.NewJoinedManagedCluster()},
-			nodes: []runtime.Object{
-				testinghelpers.NewNode("testnode1", testinghelpers.NewResourceList(32, 64), testinghelpers.NewResourceList(16, 32)),
-				testinghelpers.NewNode("testnode2", testinghelpers.NewResourceList(32, 64), testinghelpers.NewResourceList(16, 32)),
-			},
-			validateActions: func(t *testing.T, actions []clienttesting.Action) {
-				expectedCondition := metav1.Condition{
-					Type:    clusterv1.ManagedClusterConditionJoined,
-					Status:  metav1.ConditionTrue,
-					Reason:  "ManagedClusterJoined",
-					Message: "Managed cluster joined",
-				}
-				expectedStatus := clusterv1.ManagedClusterStatus{
-					Version: clusterv1.ManagedClusterVersion{
-						Kubernetes: kubeversion.Get().GitVersion,
-					},
-					Capacity: clusterv1.ResourceList{
-						clusterv1.ResourceCPU:    *resource.NewQuantity(int64(64), resource.DecimalExponent),
-						clusterv1.ResourceMemory: *resource.NewQuantity(int64(1024*1024*128), resource.BinarySI),
-					},
-					Allocatable: clusterv1.ResourceList{
-						clusterv1.ResourceCPU:    *resource.NewQuantity(int64(32), resource.DecimalExponent),
-						clusterv1.ResourceMemory: *resource.NewQuantity(int64(1024*1024*64), resource.BinarySI),
-					},
-				}
-				testinghelpers.AssertActions(t, actions, "get", "update")
-				actual := actions[1].(clienttesting.UpdateActionImpl).Object
-				testinghelpers.AssertManagedClusterCondition(t, actual.(*clusterv1.ManagedCluster).Status.Conditions, expectedCondition)
-				testinghelpers.AssertManagedClusterStatus(t, actual.(*clusterv1.ManagedCluster).Status, expectedStatus)
-			},
-		},
-		{
-			name: "merge a joined managed cluster status",
-			startingObjects: []runtime.Object{
-				testinghelpers.NewManagedClusterWithStatus(
-					corev1.ResourceList{
-						"sockets": *resource.NewQuantity(int64(1200), resource.DecimalExponent),
-						"cores":   *resource.NewQuantity(int64(128), resource.DecimalExponent),
-					},
-					testinghelpers.NewResourceList(16, 32)),
-			},
-			nodes: []runtime.Object{
-				testinghelpers.NewNode("testnode1", testinghelpers.NewResourceList(32, 64), testinghelpers.NewResourceList(16, 32)),
-				testinghelpers.NewNode("testnode2", testinghelpers.NewResourceList(32, 64), testinghelpers.NewResourceList(16, 32)),
-			},
-			validateActions: func(t *testing.T, actions []clienttesting.Action) {
-				expectedCondition := metav1.Condition{
-					Type:    clusterv1.ManagedClusterConditionJoined,
-					Status:  metav1.ConditionTrue,
-					Reason:  "ManagedClusterJoined",
-					Message: "Managed cluster joined",
-				}
-				expectedStatus := clusterv1.ManagedClusterStatus{
-					Version: clusterv1.ManagedClusterVersion{
-						Kubernetes: kubeversion.Get().GitVersion,
-					},
-					Capacity: clusterv1.ResourceList{
-						"sockets":                *resource.NewQuantity(int64(1200), resource.DecimalExponent),
-						"cores":                  *resource.NewQuantity(int64(128), resource.DecimalExponent),
-						clusterv1.ResourceCPU:    *resource.NewQuantity(int64(64), resource.DecimalExponent),
-						clusterv1.ResourceMemory: *resource.NewQuantity(int64(1024*1024*128), resource.BinarySI),
-					},
-					Allocatable: clusterv1.ResourceList{
-						clusterv1.ResourceCPU:    *resource.NewQuantity(int64(32), resource.DecimalExponent),
-						clusterv1.ResourceMemory: *resource.NewQuantity(int64(1024*1024*64), resource.BinarySI),
-					},
-				}
-				testinghelpers.AssertActions(t, actions, "get", "update")
-				actual := actions[1].(clienttesting.UpdateActionImpl).Object
-				testinghelpers.AssertManagedClusterCondition(t, actual.(*clusterv1.ManagedCluster).Status.Conditions, expectedCondition)
-				testinghelpers.AssertManagedClusterStatus(t, actual.(*clusterv1.ManagedCluster).Status, expectedStatus)
 			},
 		},
 	}
@@ -169,19 +59,10 @@ func TestSyncManagedCluster(t *testing.T) {
 				clusterStore.Add(cluster)
 			}
 
-			kubeClient := kubefake.NewSimpleClientset(c.nodes...)
-			kubeInformerFactory := kubeinformers.NewSharedInformerFactory(kubeClient, time.Minute*10)
-			nodeStore := kubeInformerFactory.Core().V1().Nodes().Informer().GetStore()
-			for _, node := range c.nodes {
-				nodeStore.Add(node)
-			}
-
 			ctrl := managedClusterJoiningController{
 				clusterName:      testinghelpers.TestManagedClusterName,
 				hubClusterClient: clusterClient,
 				hubClusterLister: clusterInformerFactory.Cluster().V1().ManagedClusters().Lister(),
-				discoveryClient:  kubeClient.Discovery(),
-				nodeLister:       kubeInformerFactory.Core().V1().Nodes().Lister(),
 			}
 
 			syncErr := ctrl.sync(context.TODO(), testinghelpers.NewFakeSyncContext(t, ""))
